+++ HISTORY.md
@@ -3,6 +3,7 @@
 
 ### Next
 
+- Added ability to run app in Docker.
 - Use native Django forms.
 - Added LESS support.
 - MVP version.